Extract portfolio upstream access
