+++ wiki/Validation-and-CI.md
@@ -298,6 +298,15 @@ contract smoke, and 1,134 unit/contract tests. Local `make ci` passed with 207 i
 1,341 combined coverage tests, 94.24% total coverage, and no known vulnerabilities after the
 governed `PYSEC-2026-161` exception.
 
+The current portfolio upstream-access extraction branch ratchets the source-file threshold to
+1,217 script-counted lines after moving cached Lotus Core, performance, and DPM upstream result
+acquisition into `portfolio_upstream_access.py`. It reduces `portfolio_service.py` from 1,237 to
+980 script-counted lines while preserving cache keys, optional-client behavior, source fan-out call
+shapes, and the public `PortfolioService` API. Focused validation passed with ruff check,
+touched-module mypy, 42 portfolio service tests, and trial refactor threshold gates proving
+`max_source_file_lines=1217` passes while `1216` fails. Full local `make check` and `make ci`
+evidence remains pending for this branch.
+
 The current portfolio transaction-summary context branch moves reporting-window resolution, YTD
 transaction pagination, defensive page-row extraction, reporting-currency fallback, and
 requested-window filtering into `portfolio_transaction_summary.py`. It reduces
